\name{mccmeiv-package}
\alias{mccmeiv-package}
\alias{mccmeiv}
\docType{package}
\title{
\packageTitle{mccmeiv}
}
\description{
\packageDescription{mccmeiv}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{mccmeiv}
\packageIndices{mccmeiv}
This package implements the methodology found in Manuel et al. to estimate parameters of a logistic model using data from a 
matched case control design with a mismeasured exposure variable by using instrumental variables.
}
\author{
\packageAuthor{mccmeiv}

Maintainer: \packageMaintainer{mccmeiv}
}

\references{
Manuel, CM, Sinha, S and  Wang, S. Matched case-control data with a misclassified exposure: 
What can be done with instrumental variables? (Submitted).
}

\keyword{ Mismeasured}
\keyword{ Matched Cased Control}
\keyword{ Instrumental}
\keyword{ bias}



