\name{mccmeiv-package}
\alias{mccmeiv-package}
\alias{mccmeiv}
\docType{package}
\title{
\packageTitle{mccmeiv}
}
\description{
\packageDescription{mccmeiv}
}
\details{
This package implements the methodology found in Manuel, Wang and Sinha (2018) used to estimate parameters of a matched case control design
with a mismeasured exposure variable by using instrumental variables.
}
\author{
\packageAuthor{mccmeiv}

Maintainer: \packageMaintainer{mccmeiv}
}

\references{
Manuel, CM, Wang, S, and Sinha, S (2018). Matched case-control data with a misclassified exposure: 
What can be done with instrumental variables? 
}

\seealso{
\code{\link{meiv}}
\code{\link{matcdata}}
\code{\link{matcdata2}}
\code{\link{matcdatamult}}
}

\keyword{ Mismeasured}
\keyword{ Matched Cased Control}
\keyword{ Instrumental}
\keyword{ bias}

